carry more layout structs to `data` crate
