+++ data/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod charts;
+pub mod layout;
 pub mod pane;
